alx_travel_app/requirement.txt
celery==5.4.0
Django==4.2.17
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
drf-yasg==1.21.8
rabbitmq==0.2.0